tests/t-kdf: fix format string argument type mismatch
* tests/t-kdf.c (check_pbkdf2, check_scrypt): Change fprintf format
for plen from "%ld" to "%u" and cast plen to 'unsigned int'.
--

Fix compiler warning: "format ‘%ld’ expects argument of type ‘long int’, but
argument 3 has type ‘size_t’ {aka ‘unsigned int’}".

+++ tests/t-kdf.c
@@ -1118,8 +1118,8 @@ check_pbkdf2 (void)
             {
               if (verbose)
                 fprintf (stderr,
-                         "  shorter key (%ld) rejected correctly in fips mode\n",
-                         tv[tvidx].plen);
+                         "  shorter key (%u) rejected correctly in fips mode\n",
+                         (unsigned int)tv[tvidx].plen);
             }
           else
             fail ("pbkdf2 test %d failed: %s\n", tvidx, gpg_strerror (err));
@@ -1225,8 +1225,8 @@ check_scrypt (void)
             {
               if (verbose)
                 fprintf (stderr,
-                         "  shorter key (%ld) rejected correctly in fips mode\n",
-                         tv[tvidx].plen);
+                         "  shorter key (%u) rejected correctly in fips mode\n",
+                         (unsigned int)tv[tvidx].plen);
             }
           else
             fail ("scrypt test %d failed: %s\n", tvidx, gpg_strerror (err));
